+++ R/Drivers.R
@@ -28,7 +28,7 @@ jdbcDrivers <- new.env()
 #' - "redshift" for Amazon Redshift
 #' - "sql server", "pdw" or "synapse" for Microsoft SQL Server
 #' - "oracle" for Oracle
-#' - "spark" for Spark
+#' - "spark" for Spark (Databricks)
 #' - "snowflake" for Snowflake
 #' - "bigquery" for Google BigQuery
 #' - "iris" for InterSystems IRIS
@@ -41,13 +41,13 @@ jdbcDrivers <- new.env()
 #' The following versions of the JDBC drivers are currently used:
 #' 
 #' - PostgreSQL: V42.7.3
-#' - RedShift: V2.1.0.9
+#' - RedShift: V2.1.0.20
 #' - SQL Server: V9.2.0
 #' - Oracle: V19.8
-#' - Spark (Databricks): V2.6.36
-#' - Snowflake: V3.24.0
-#' - BigQuery: v1.3.2.1003
-#' - InterSystems IRIS: v3.10.2
+#' - Spark (Databricks): V3.3.3
+#' - Snowflake: V4.2.0
+#' - BigQuery: v0.10.0
+#' - InterSystems IRIS: v3.11.0
 #' 
 #' @return Invisibly returns the destination if the download was successful.
 #' @export
@@ -94,10 +94,10 @@ downloadJdbcDrivers <- function(dbms, pathToDriver = Sys.getenv("DATABASECONNECT
     2,redshift,redshift-jdbc42-2.1.0.20.zip,https://s3.amazonaws.com/redshift-downloads/drivers/jdbc/2.1.0.20/
     3,sql server,sqlServerV9.2.0.zip,https://ohdsi.github.io/DatabaseConnectorJars/
     4,oracle,oracleV19.8.zip,https://ohdsi.github.io/DatabaseConnectorJars/
-    5,spark,DatabricksJDBC42-2.6.36.1062.zip,https://databricks-bi-artifacts.s3.us-east-2.amazonaws.com/simbaspark-drivers/jdbc/2.6.36/
-    6,snowflake,snowflake-jdbc-3.24.0.jar,https://repo1.maven.org/maven2/net/snowflake/snowflake-jdbc/3.24.0/
-    7,bigquery,SimbaJDBCDriverforGoogleBigQuery42_1.6.2.1003.zip,https://storage.googleapis.com/simba-bq-release/jdbc/
-    8,iris,intersystems-jdbc-3.10.2.jar,https://repo1.maven.org/maven2/com/intersystems/intersystems-jdbc/3.10.2/"
+    5,spark,DatabricksJDBC-3.3.3.zip,https://databricks-bi-artifacts.s3.us-east-2.amazonaws.com/simbaspark-drivers/jdbc/3.3.3/
+    6,snowflake,snowflake-jdbc-4.2.0.jar,https://repo1.maven.org/maven2/net/snowflake/snowflake-jdbc/4.2.0/
+    7,bigquery,google-cloud-bigquery-jdbc-0.10.0-all.jar,https://repo1.maven.org/maven2/com/google/cloud/google-cloud-bigquery-jdbc/0.10.0/
+    8,iris,intersystems-jdbc-3.11.0.jar,https://repo1.maven.org/maven2/com/intersystems/intersystems-jdbc/3.11.0/"
   )
 
   if (dbms == "all") {