+++ inst/sql/sql_server/sparkCopy.sql
@@ -1,10 +1,10 @@
 COPY INTO @sqlTableName
-FROM 'abfss://@azureStorageAccount.dfs.core.windows.net/@fileName'
-WITH (
- CREDENTIAL (AZURE_SAS_TOKEN = '@azureAccountKey')
-)
+FROM (
+  SELECT @selectFields
+  FROM 'abfss://@azureContainerName@@azureStorageAccount.dfs.core.windows.net/@fileName'
+) 
 FILEFORMAT = CSV
 FORMAT_OPTIONS (
-   'header' = 'true',
-   'inferSchema' = 'true'
-);
+   'header' = 'true'
+)
+COPY_OPTIONS('mergeSchema' = 'true');